Adding deleteTopic and move the styles
