+++ src/screen/topic/TopicView.js
@@ -6,7 +6,6 @@ import {
     View,
     Text,
     Image,
-    StyleSheet, 
     Dimensions, 
     ScrollView, 
     FlatList,
@@ -20,7 +19,7 @@ import {
 } from '@expo/vector-icons';
 import Header from './header'
 import { TextInput } from 'react-native-gesture-handler';
-
+import styles from './styles';
 import AvatarUser from '../../assets/AvatarUser.png';
 
 const largura = Dimensions.get("screen").width;
@@ -28,14 +27,15 @@ const altura = Dimensions.get("screen").height;
 
 export default function TopicView(){
 
-    const enderecoIpv4 = '192.168.18.7'; //inserir o endereço o ip do localhost
+    const enderecoIpv4 = '192.168.0.40'; //inserir o endereço o ip do localhost
     const porta = '3000'; // inserir a porta em que o backend esta rodando
-    const topicID = '5fa56d06904f9400289e158a'; //inserir o id da planta  ser exibido
+    const topicID = '5fa473db6ad7aa001cb7654a'; //inserir o id da planta  ser exibido
     const [topic, setTopic] = useState({})
     const [isNewComment, setIsNewComment] = useState(true);
     const [isEditing, setIsEditing] = useState(false);
     const [commentText,setCommentText] = useState('');
     const [isLiked, setIsLiked] = useState(false);
+    const [isDeletd, setIsDeletd] = useState(false);
 
     useEffect(() => {
         fetch(`http://${enderecoIpv4}:${porta}/topic/find/${topicID}`)
@@ -45,7 +45,7 @@ export default function TopicView(){
       }, []);
     
     const putTopic = async () => {
-      const requestOptions = {
+        const requestOptions = {
           method:'PUT',
           headers: {
               'Content-Type':'application/json'
@@ -56,9 +56,26 @@ export default function TopicView(){
               })
           };
           fetch(`http://${enderecoIpv4}:${porta}/topic/update/${topic._id}`,requestOptions)
-              .then(response => response.json())
-              .then(() => setIsEditing(!isEditing))
-              .catch((error) => console.error(error))         
+            .then(response => response.json())
+            .then(() => setIsEditing(false))
+            .catch((error) => console.error(error))         
+}
+    const deleteTopic = async () => {
+        const requestOptions = {
+        method:'PUT',
+        headers: {
+            'Content-Type':'application/json'
+        },
+        body:JSON.stringify({
+            title:'Esse tópico foi deletado',
+            description:'Este tópico foi deletado'
+            })
+        };
+        fetch(`http://${enderecoIpv4}:${porta}/topic/update/${topic._id}`,requestOptions)
+            .then(response => response.json())
+            .then(data => setTopic(data))
+            .then(() => setIsDeletd(true))
+            .catch((error) => console.error(error))         
     }
 
     const postComment = async () => {
@@ -178,28 +195,32 @@ export default function TopicView(){
                     </View>
                     <View style={styles.topicDiv}>
                         {
-                        !isEditing?
-                            <>
-                                <View style={{marginTop:10}}>
-                                    <Text style={styles.topicDivTitle}>{topic?.title}</Text>
-                                    </View>
-                                <ScrollView style={styles.scrollViewDescription}>
-                                    <Text style={styles.topicDivDescription}>{topic?.description}</Text>
-                                    {!!topic.profile_picture && 
-                                        <Image
-                                            style={styles.imgDescription}
-                                            source={{uri: topic.profile_picture}}
-                                        />
-                                    }   
-                                </ScrollView>
-                            </>:
+                        isEditing?
                             <>
                             <View style={{marginTop:10}}>
                                 <TextInput style={styles.topicDivTitle} defaultValue={topic?.title} onChangeText={(title) => setTopic({...topic,title})}></TextInput>
                                 </View>
                                 <TextInput blurOnSubmit multiline style={styles.topicDescriptionInput} defaultValue={topic?.description} onChangeText={(description) => setTopic({...topic,description})}></TextInput>
+                                {!isDeletd &&
                                 <TouchableOpacity style={styles.saveButton} onPress={() => putTopic()}><Text style={styles.saveButtonText}>salvar</Text></TouchableOpacity>
-                        </>
+                                }
+                                </>: 
+                            <>
+                            <View style={{marginTop:10}}>
+                                <Text style={styles.topicDivTitle}>{topic?.title}</Text>
+                                </View>
+                            <ScrollView style={styles.scrollViewDescription}>
+                                <Text style={styles.topicDivDescription}>{topic?.description}</Text>
+                                {!!topic.profile_picture && 
+                                    <Image
+                                        style={styles.imgDescription}
+                                        source={{uri: topic.profile_picture}}
+                                    />
+                                }   
+                            </ScrollView>
+                            <Feather style={styles.deleteButton} onPress={() => deleteTopic()} name="trash-2" size={24} color="black" />
+                            
+                            </>
                         }
                         <View style={styles.topicContainer} >
                             <View style={styles.topicDivLikes}>
@@ -207,7 +228,7 @@ export default function TopicView(){
                                         <Text>{topic.likes - topic.dislikes}</Text>
                             </View>
                             <View style={styles.commentIcon}>
-                                <Feather name="edit" size={18} color="black" onPress={() => setIsEditing(!isEditing) }/>
+                                <Feather name="edit" size={18} color="black" onPress={() => !isDeletd && setIsEditing(!isEditing) }/>
                             </View>
                             <View style={styles.shareIcon}>
                                 <Feather name="corner-up-right" size={15} color="black" />
@@ -245,199 +266,3 @@ export default function TopicView(){
         </KeyboardAvoidingView>        
       );
   };
-const styles = StyleSheet.create({
-    UserDiv: {
-        flexDirection: 'row',
-        flex:0.6,
-        marginTop:altura/50
-        
-    },
-    imgUser : {
-        width:largura/5,
-        height:largura/5,
-        borderRadius:50,
-    },
-    nameUser : {
-        fontSize:18,
-        fontWeight:'400',
-        paddingTop:5,
-        paddingLeft:10
-        
-    },
-    dataTopic : {
-        fontSize:12,
-        fontWeight: '200',
-        paddingLeft:10,
-        paddingTop:0
-    },
-    topicDiv:{
-        flexDirection: 'column',
-        alignContent:'space-between',
-        flex:1.4,        
-        
-    },
-    topicDivTitle:{
-        fontSize:18,
-        fontWeight:'600'
-    },
-    scrollViewDescription:{
-        maxHeight:altura/4.4,
-    },
-    imgDescription:{
-        alignSelf:"center",
-        width:largura/1.2,
-        height:largura/1.2,
-        margin:10
-    },
-    topicDescriptionInput:{
-        flex:1
-    },
-    saveButton:{
-        backgroundColor:'#D8A3E0',
-        alignSelf:'flex-end',
-        justifyContent:'center',
-        margin:10,
-        height:largura/12,
-        width:largura/4,
-        borderRadius:10
-    },
-    saveButtonText:{
-        fontSize:16,
-        color:'black',
-        textAlign:'center',
-        textAlignVertical:"center",
-        
-    },
-
-    topicDivDescription:{
-        fontSize:12,
-        fontWeight:'400',
-        paddingTop:20
-    },
-    topicContainer:{
-        flexDirection: 'row',
-        flex:6,
-        backgroundColor:'#D6DADF',
-        maxHeight:altura/20,
-        borderRadius:8,
-        alignItems:'center',
-        justifyContent: 'space-between',
-        
-    },
-    topicDivLikes:{
-        flexDirection: 'row',
-        alignItems:'center',
-        justifyContent:'flex-start',
-        paddingLeft:10
-        
-    },
-    shareIcon : {
-        flexDirection: 'row',
-        alignItems:'center',
-        justifyContent: 'flex-end',
-        paddingHorizontal:10,
-        
-    },
-    commentsBarDiv:{
-        flex:1,
-        marginTop:20,
-        backgroundColor:'white'
-    },
-    commentsListDiv:{
-        flex:5,
-        backgroundColor:'#D8A3E0',
-        justifyContent: 'center',
-        flexWrap:'nowrap',
-    },
-    commentsBar:{
-        flex:5,
-        backgroundColor:'#D8A3E0',
-        alignContent:'flex-start',
-        flexDirection:"row",
-    },
-    commentsBarText:{
-        fontSize:16,
-        fontWeight:'500',
-        alignSelf:'flex-start',
-        paddingLeft:largura/10,
-        paddingTop:12        
-    },
-    commentsBarIcon:{
-        alignSelf:'center',
-        justifyContent:'center',
-        paddingLeft:5 
-    },
-    commentsList:{
-        flex:1,
-        backgroundColor:'#D6DADF'
-    },
-    commentItemDiv:{
-        backgroundColor:'white',
-        flexDirection:'row',
-        alignSelf:'center',
-        width:largura/1.1,
-        height:altura/8,
-        marginVertical:5,
-        borderRadius:5,
-        paddingTop:10,
-        paddingLeft:10
-    },
-    imgUserComment:{
-        marginLeft:10,
-        width:largura/8,
-        height:largura/8,
-        borderRadius:50,
-    },
-    commentUsername:{
-        fontSize:12,
-        fontWeight:'bold',
-        paddingTop:5,
-        paddingLeft:10
-    },
-    commentData:{
-        fontSize:7,
-        fontWeight:'200',
-        paddingTop:5,
-        paddingLeft:10
-    },
-    commentDescription:{
-        fontSize:9,
-        fontWeight:'300',
-        paddingTop:5,
-        paddingLeft:10
-    },
-    container: {
-        flex:7,
-        flexDirection: 'column',
-        justifyContent: 'center',
-        flexWrap:'nowrap' ,
-        width: largura/1.2,
-        height:altura,
-        alignSelf:'center'
-        
-    },
-    containerMaster:{
-        flex:1,
-        flexDirection: 'column',
-        justifyContent: 'center',
-        flexWrap:'nowrap' ,
-        width:largura,
-        height:altura,
-    },
-    commentContent:{
-        flexDirection:'column',
-        alignContent:'center',
-        alignContent:'center',
-        alignSelf:'center',
-        padding:15
-    },
-    commentDescriptionInput:{
-        backgroundColor:'white',
-        height:altura/5,
-        width:largura/1.2,
-        alignSelf:'center',
-        borderRadius:5,
-        padding:10,
-
-    }
-})
